Renaming provider to collector
