+++ Tests/Unit/Provider/TaskStatusProviderTest.php
@@ -8,7 +8,7 @@
  * @license     GNU/GPLv3 http://www.gnu.org/licenses/gpl-3.0.html
  */
 
-namespace MauticPlugin\CronfigBundle\Tests\Unit\Model;
+namespace MauticPlugin\CronfigBundle\Tests\Unit\Provider;
 
 use Mautic\CoreBundle\Helper\CoreParametersHelper;
 use PHPUnit\Framework\MockObject\MockObject;
